+++ crates/ark/src/analysis/mod.rs
@@ -5,4 +5,4 @@
 //
 //
 
-pub mod parse_boundaries;
+pub mod input_boundaries;
